+++ src/main/java/com/project/momo/entity/MemberPointsHistory.java
@@ -33,14 +33,14 @@ public class MemberPointsHistory extends BaseEntity {
     @NotBlank
     private String description;
 
-    private int amount;
+    private long amount;
 
     @Column(name = "points_before")
     @Min(0)
-    private int pointsBefore;
+    private long pointsBefore;
 
     @Column(name = "points_after")
     @Min(0)
-    private int pointsAfter;
+    private long pointsAfter;
 
 }
